refactor: file structure of components
